refactor(money): UI'da kuruş gösterimi + input dönüşümü
format.ts formatTRY artık kuruş kabul ediyor (money.formatKurus'a delege; tek
formatter sözleşmesi). Tüm para gösterimleri formatTRY(kuruş) üzerinden:
- Ham formatlamalar kaldırıldı: whatsapp paylaşım metni, send-reminder-button,
  parts/format formatPrice (minor units /100), technician-order-detail kalem &
  toplam satırları.
- Oran gösterimleri bpsToPercent ile: technician/quote-detail KDV etiketi,
  customer-detail iskonto.

Fiyat inputları TRY ondalık alıp gönderimde kuruşa çevriliyor (money modülü);
oran inputları yüzde alıp bps'e çevriliyor:
- quote-create-form: katalog seçim kuruş->TRY, önizleme calculateOrderTotals ile
  kuruş (server'la aynı), submit'te TRY->kuruş + %->bps; client grandTotal artık
  gönderilmiyor (server otoriter).
- order-detail: kalem ekleme fiyatı, iskonto/KDV meta draft (kuruş/bps), önizleme
  money modülüyle.
- customer-create-form (discountRate %<->bps), part-form (fiyat TRY<->kuruş),
  collection-create-form (tutar TRY->kuruş, önizleme kuruş).
- quote-detail toplamları calculateOrderTotals ile (inline float math kaldırıldı).

analytics/queries ve dashboard/queries: (subtotal * taxRate)/100 -> applyTaxBps
(taxRate bps) + addKurus. Form (customer/part) şemaları TRY/% girişini kabul
edecek şekilde tutuldu; server şemaları kuruş/bps doğruluyor.

+++ src/lib/validations/part.ts
@@ -10,8 +10,8 @@ export const partSchema = z.object({
   unit: z.string().min(1, "Birim zorunludur").default("adet"),
   stockQty: z.coerce.number().min(0).default(0),
   criticalStockQty: z.coerce.number().min(0).default(0),
-  purchasePrice: z.coerce.number().int().min(0).optional().default(0), // kuruş
-  salePrice: z.coerce.number().int().min(0).optional().default(0), // kuruş
+  purchasePrice: z.coerce.number().min(0).optional().default(0), // TRY in the form; converted to kuruş on submit
+  salePrice: z.coerce.number().min(0).optional().default(0), // TRY in the form; converted to kuruş on submit
   currency: z.enum(["TRY", "USD", "EUR"]).default("TRY"),
   supplierName: z.string().optional().default(""),
   supplierPhone: z.string().optional().default(""),
